postgresql: enable global listening, and explicit IP permissions
To allow better out of the box configuration of postgresql, two things
are required: listen on all addresses, and add explicit permissions for
the known nodes in the network.

In this change, the former is done by adding listen_addresses = '*' to the
generated postgresql configuration on first boot, as well as adding the
known compute and controller IPs to pg_hba.conf. (note: the values for these
IPs come from hosts.bbclass)

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>