openstack: extra space, network facilities to controller image and deployment
Various components (such as rabbitmq) have checks for minimum amounts
of free space on a image. If those checks fail, the service exits in
a hard to debug manner. To ensure that these services run, and meet
typical usage scenarios, we increase the free space in the image to 3G.

We also automatically include the network node packages into this image
type, to ensure that full quantum functionality is available.

Finally, we inherit openstack-image-base, to trigger simple rootfs
configuration and deployment.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>