postgresql: Fix init to work with postgresql in a drbd data container
When postgresql installed on the target and configured with drbd the
data will live in the drbd container and the startup script must
account for the different invocation to pg_ctl, as well as checking
that the config files are properly linked into to the postgresql data
directory.

The set -e was removed because the script was exiting improperly
during the restart operation and status operations.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>