tempest: create per-component test packages and flakes8
To add more complete tempest support, we require flakes8, so it is
added to the dependency list.

To get the individual component test scripts onto the target, create
a $PACKAGE-tests package and add the script. When the tests are
required on target, these packages should be added to the install
list.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>