python-*: prevent setuptools from fetching required eggs
Several python packages require 'python-pbr' both at build and
runtime, as listed in their respective setup.py files, yet this
dependency is not included in their recipe. Adding python-pbr
to the RDEPENDS to correct this.

In addition this situation is complicated by the fact that the
setuptools will actually fetch python-pip and python-pbr eggs,
regardless of the value of BB_NO_NETWORK, if any of these packages are
built before python-pip and python-pbr are in the sysroot. Most
dramitically if you were to attempt to build any of these packages
with no network connectivity the do_compile() task will fail with the
following:

| DEBUG: Executing shell function do_compile
| Download error: [Errno 110] Connection timed out -- Some packages may not be found!
| Couldn't find index page for 'pip' (maybe misspelled?)
| Download error: [Errno 110] Connection timed out -- Some packages may not be found!
| No local packages or download links found for pip>=1.0
| Traceback (most recent call last):
|   File "setup.py", line 21, in <module>
|     pbr=True)

Adding the missing DEPENDS will ensure these packages are available
without the need for setuptools to fetch them, and avoid possible
build issues due to network connectivity.

In order to test these modifications all of these packages have been
built with a populated sstate cache and the network crippled using:

iptables -A OUTPUT -p tcp --destination-port 80 -j DROP

to ensure no extra fetches are taking place.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>