OpenStack: nova-novncproxy:can't stop this service
nova-novncproxy and nova-spicehtml5proxy both do not check to see if the process
specified in PIDFILE exists before firing off a new one. Both initrd files can
actually use nova.init (used for the many other nova-* daemons)

nova-consoleauth also has this issue, we add a check in start to see if the process
in PIDFILE exists and exits if it does.

Signed-off-by: Amy Fong <amy.fong@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>