python-neutron: remove greedy FILES_
We need to use specific filenames instead of a glob to prevent the
linuxbridge and openvswitch ini files from being bundled as part of
the main package as opposed to their respective packages. Without this
change the FILES_ rules for the sub-packages are not effective and the
resulting packages are not being populated as expected.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>