OpenStack: modify controller to be able to talk to both console
We modify the controller so that both vnc and spice html5 proxies are running.
Since in the current recipe, both controller and compute share the same nova.conf
file, in the compute, both vnc and spice html5 are set to enabled and in this
configuration, vnc is chosen (in source: vnc is chosen because it's better
tested).

To change to spice, only change required is on the compute side.

The controller (since the default console_type is auto) will try to talk
to both and run whichever is available.

Signed-off-by: Amy Fong <amy.fong@windriver.com>