spice: updates to work with latest autotools
We were getting several errors during configure related to
autoreconf. We need to create an empty m4 directory as well
as enable the subdir-objects option to allow configure to
complete successfully. There were indications that this second
issue was fixed upstream but even using the latest git trees
I am seeing this failure.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>