Memory leak from timings
Memory leak happens when the dynamic list times grows without anything to
reset it.

In ceilometer and cinder configuration files, the new option is created:
[nova_client]
max_timing_buffer=<value>

In all clients found that uses extends the HTTPClient and uses the times
list, we limit the size of the list by popping off the oldest item in the
list to maintain a maximum size. A default size of 200 is chosen,
configurable by the above configuration option.

Signed-off-by: Amy Fong <amy.fong@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>