python-neutron: openvswitch plugin read config files
Currently the openvswitch plugin doesn't read any config files so we
are unable to configure it properly. Have the init script pass in the
config files we are already installing.

The config needs local_ip set otherwise it will fail to run.

We can't just tack rabbit_host on the end of the conf file as it is
in the wrong section, so change this to a substitution.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>