barbican: remove unneeded files from being packaged
Issue: US-34303

Barbican source code comes with scripts that are intended to control
the service.  Added previously was a script for this same purpose
that is placed into init.d that integrates more consistently with
the system.  This makes the need for these scripts redundant.  This
patch removes the scripts being put into the final system package.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>