barbican: increase uwsgi buffer-size to work with keystone
uWSGI defaults to a maximum packet size of 4096 bytes.  This
is too small to support working with PKI tokens that are now default
in Keystone.  The size of the packets within Barbican are dependent
on both the size of the Keystone token and the size of the secret to
be stored & retrieved.  Increasing the buffer size to the maximum
allowed by uWSGI allows Barbican to support the largest possible
secrets.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>