sysklogd: add syslog user and group and add dependency
Some OpenStack components expect the system to contain a
syslog user for logging purposes and fail if one doesn't
exist. This fix adds the syslog user and group within the
sysklogd package since it responsible for the system
logging capabilities of the system.  This fix also adds a
dependency between the package requiring the syslog user
to exist.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>