horizon: patch test script to run without coverage
Currently, the run_tests.sh script provided by the horizon
package requires the command "coverage" to be available
even if the "-c" option (test with code coverage) is not
specified on the command line.  This fix patches the test
script to remove the calls to the "coverage" tool and calls
the test script directly if the "-c" option is not provided.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>