horizon: add full path to test directories
The tests included Horizon expect all files to be in a single
location in order to be run directly from the source tree. The
recipe for installing Horizon on the system installs Horizon
as a python site-package and puts files in different locations
depending on the file type.  In order to have the tests support
this type of install we need to explicitly indicate the full path
of the test files to exclude in order to have the tests run
successfully.  This fix adds the absolute path to the test
locations allows the Horizon tests to pass as expected.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>