keystone: install example test certificates
Some tests provided by Keystone tests signing with an
example certificate and signing key.  If these certificates
are not found these particular tests will hang.  Thus, in
order for these tests to pass we must install the example
certificates to the system.  This fix updates the install
script for Keystone to include installing the example
certificates.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>