keystone: fix paths for testing
Keystone tests are designed to run on the source tree.
However, Keystone is installed on a system with files
in various directories.  This fix patches the testing
source files to be able to find the files on the
distribution.  This fix incorporates the changes of
a previous patch file into a new patch file that is
generated, since the previous patch are related and
close to eachother in the source and it is easier to
maintain less patch files.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>