cloud-init: install to correct target directory
The setup.py script that is included with the cloud-init code
is setup to install files into the "/usr/lib" directory.  On
64-bit machines the target directory should be "/usr/lib64".
This generates a warning when using bitbake.  This fix performs
a search and replace on the setup.py file to set the correct
output directory.  This fix also adds a symbolic link to each
script from the /etc/cloud directory to make the scripts
available from a more obvious location.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>