nova: package doc directory and fix its path
The tests included with nova require some setup and configuration
files that are within the "doc" directory.  This fix includes the doc
directory on the target system as part of the recipe.  This fix also
includes a patch that updates the tests to find the "doc" directory
where it is installed on the target system.

Signed-off-by: Keith Holman <Keith.Holman@windriver.com>