cinder/glance/neutron: support for openstackchef class
Openstackchef enables us to recreate configuration files
for services in an openstack installation. It does this by
creating template file(s) out of openstack services' configuration
file(s) exposed to it.

The following services are inheriting the openstack class
and then exposing a set of configuration files to the class.
The configuration files are assumed to have been installed in the image
in the service's WORKDIR and are exposed to openstackchef by assigning
them to the variable CHEF_SERVICES_CONF_FILES.

For legacy reasons, the string OPENSTACKCHEF_ENABLED is defined in
openstackchef class, but it can be overwritten to an empty string in
a .bb, .class, .bbapend or local.conf file when openstackchef support
is not desired. This enables all of these services to be built without
openstackchef support. In addition, it prevents the recipes for these
sevices from substituting the placeholders in their configuration
files when inheriting openstackchef.

However, since services like python-glance and python-cinder install some
of their configuration files to the image directory during the
substitution, it means that these files at not installed when the
substitutions are not required.
Therefore, we have taken the liberty to install those missing configuration
files before the check for whether or not substitutions should be made,
without having any side effect on when the substitutions are required.

At build-time, openstackchef makes chef-solo templates out of
the registered files. And at run-time, the deploychef package
makes a call to chef-solo, which in-turn use the template files
to recreate the registered configuration files.

In making template files out of configuration files, openstackchef
makes a simple placeholder/value substitution in the configuration files.
However, it provides a mechanism for services inheriting the class
to define a special shell callback function which can do more than
a simple placeholder/value substitution. This special callback
function is exposed to openstackchef by assigning it to the variable
CHEF_SERVICES_SPECIAL_FUNC. And the name of the file passed back to the shell
callback function as an argument is set to the variable
CHEF_SERVICES_FILE_NAME by openstackchef. In this patch set, only
python-neutron and python-glance recipe files makes special
substitutions.

Signed-off-by: Mustapha Lansana <Mustapha.Lansana@windriver.com>