Keystone: package service/user addition sstate vardeps
Several packages utilize the keystone package service/user addition
services recently added. The data passed to this service depends on
the value assigned to CONTROLLER_IP (used as KEYSTONE_HOST), however,
bitbake is not able to automatically determine this dependency so
several tasks which should be rerun to create updated package postinst
scripts are not run when CONTROLLER_IP is modified. Adding the
necessary vardeps ensure these tasks are rerun and now adjustments
made to CONTROLLER_IP are reflected in the generated packages.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>