dhcp: fix double 'request' in dhclient.conf
commit 43e2b093040734d4c4b28db96a634cd2a5f3271b [dhcp: add dhcp
classless static route support] had a flaw in one of
do_install_append() sed expressions in the dhcp recipe that caused a
duplicate 'request' keyword to be inserted in to the dhclient.conf
file. The result was that results returned for the request were not as
expected and the resolv.conf would be empty. Fixing up the sed
expression removes the duplicate and allows things to function
properly.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>