python: remove recipes that are now in meta-virtualization
In order for meta-virtualization to include a recipe for docker and
docker-registry these recipes have been moved to allow
meta-virtualization to exist without a circular dependency with
meta-cloud-services. There is already a dependency here on
meta-virtualization so this will ensure recipes in meta-cloud-services
will still build as they did before.

A concern does exist that several OpenStack recipes have a tight
binding on specific versions of these recipes. This move doesn't
prevent meta-cloud-services from hosting its own version of these
recipes in the future and making use of 'preferred' version to control
this binding.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>