meta-openstack: remove unneeded dangling bbappends
Several bbappends are now dangling as the base recipes have moved to
different versions. These specific danglers have been reviewed and
don't have any demonstrable purpose any longer and can therefor be
removed.

For example the busybox default config now enables CONFIG_WHICH so
there is no reason to continue to 'patch' this config setting.

For recipes which used brokensep in order to work around S != B build
issues the fact that the bbappend was dangling and the recipe was
building successfully is enough to demonstrate that the bbappend was
no longer needed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>