dnsmasq: move bbappend to use a wilcard
To facilitate future support and to simplify using meta-cloud-services
with other branches we prefer to use a wildcard. In making this move
the content of the bbappend has been audited and most of the content
is now no longer needed (it is found in dnsmasq.inc for some time now)
and so we are removing these parts.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>