python packages: add build dependency on python-pbr
Some packages fail to build because their dependency on
python-pbr during built time is not properly stated in the
recipe. Sometimes the build succeeds anyway because the
python-pbr package has already been built previously. To
avoid the occasional build failure, this patch adds a
dependency on python-pbr to all those packages that
declare it as a dependency in their setup.py file.

Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>