python-django: uprev to 1.8.6
Move to use git to ease future uprevs. Note that a recipe for django
exists in meta-python and since it is a versioned recipe instead of a
"_git" recipe it seems to be picked up instead of our recipe, so we
need to add a PREFERRED_VERSION to ensure our recipe is used.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>