qemu: Add packageconfig for libseccomp to avoid rdepend QA warning
"WARNING: QA Issue: qemu rdepends on libseccomp, but it isn't a build
dependency? [build-deps]"

Currently qemu itself checks and links with libsecomp(if version >= 2.1.1)
for x86 and x86-64, since it is not specified in configure, and causes the
above warning.

This patch adds packageconfig for libseccomp and set it disabled as default.
This feature can be enabled whenever libseccomp exists and is needed.

Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>