python: satisfy setup.py 'setup_requires'
Python setuptools will attempt to satisfy the packages defined as
'setup_requires' in setup.py by first looking for the package
availability locally and ultimately by downloading it from PyPI. This
is actually a huge security hole and packages should move to using pip
instead, but this is another story that the upstream packages have to
address. This also disregards BB_NO_NETWORK and may prove to introduce
host contamination.

The best approach is to ensure we have the -native version of the
'setup_requires' packages present such that setup.py will not attempt
to complete the download from PyPI.

Make 'pbr' -native available and for packages which we have identified
as having 'setup_requires' include 'pbr' add the necessary
python-pbr-native DEPENDS.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>