python-positional: use archived tarball
The upstream project has been discontinued as you can see from the
comment from the developer in response to another pkg which used
positional:
https://github.com/conda-forge/positional-feedstock/issues/7

In addition to discontinuing the project the developer removed the
repository, so we need to use the version archived via pypi. We can't
simply drop the package as it is used by python-nova and other
openstack components.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>