rabbitmq-server: add missing depends on coreutils-native
With the move to RSS this missing dependency became a build issue,
previously hidden by some other package fullfilling the dependency
before this package was built. With RSS we need to ensure all first
level dependencies are listed to avoid build failures. The build
requires the 'fold' command which is provided by coreutils-native.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>