python-*: switch remaining pypi recipes to use the pypi class
Cleanup the various python-* recipes which download from pypi. The
biggest change is to "inherit pypi" which should result in us always
using current pypi best practices. This will for example ensure we are
using https and not http which is apparently going to be disabled
soon.

In most cases we are able to drop SRC_URI, however, for some recipes
using archived versions of packages we must set PYPI_SRC_URI to
overwrite what the pypi class would derive. For example when the
archive is a zip and not a tar.gz.

In all cases we can drop the defining of 'S' and 'PR' as the pypi
class and the PR server will set these correctly.

In most cases we can drop SRCNAME. Where needed we instead set
PYPI_PACKAGE to overwrite the derived name that the pypi class would
otherwise calculate.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>