python-jsonschema: drop our recipe in favor of the one in meta-python
The current v2.5.1 recipe is slightly broken (needs PYPI_PACKAGE_EXT =
"zip" added). But regardless the current v2.5.1 recipe is unused since
we are picking up the v2.6.0 in meta-python. Drop the unused/broken
v2.5.1 recipe to clean things up. As with all our python recipes we
will need to keep an eye on the version to ensure the are compatible,
so add a PREF version to help track any skew that might develop.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>