postgresql: updates to get things working with systemd
Convert the sysvinit code to instead work with systemd. We are no
longer able to make use of postinst scripts as we need to perform
setup after the postgresql service is started, this will have
'knock-on' effects for other postinst, such as for keystone. Changing
these postinst to "one time" services buys us greater control and
easier readability than the postinst scripts, so overall this is a
good change.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>