python-keystone: generate admin-openrc
In order to interact with the openstack CLI commands it is common
practice to source *openrc files which populate the environment with a
base set of values. See

https://docs.openstack.org/keystone/pike/install/keystone-openrc-obs.html

We used to do this as part of nova but it makes more sense to create
these as part of keystone. This makes them available early in the
setup process and also maps with the keystone bootstrapping.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>