python-glance: uprev to latest stable/pike
Complete the uprev and the transition to systemd. The configuration is
cleaned up to match the setup documentation found in the openstack
installation guide. Once a build is completed with these changes the
controller image is able to have images added via:

 %> . /etc/keystone/admin-openrc
 %> openstack image create "cirros" \
   --file cirros-0.3.5-x86_64-disk.img \
   --disk-format qcow2 --container-format bare \
   --public

Images can be listed with 'openstack image list' or
'glance image-list' or via the horizon web interface.

Requires introduction of python-glance-store...

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>