libibverbs: don't do battle when meta-dpdk is in play
The meta-dpdk layer has its own special version of libibverbs which
will conflict with our version:

  Error: Transaction check error:
  file /usr/lib64/libibverbs.so.1.0.0 conflicts between attempted installs of \
  libibverbs1-1.1.7-r0.corei7_64 and dpdk-dev-libibverbs-1.2.1+3.4+2.0.0.0-r0.corei7_64

Ensure only our version or the meta-dpdk version is in use. With
current layer priorities and such this should result in the meta-dpdk
version being used, which is fine since if DPDK is desired we are best
using the version recommended by meta-dpdk. Users can use
PREFERRED_VERSION or other tools to override the default behavior.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>