python-ansible: copy recipe from meta-overc (branch:master-oci)
This recipe was part of meta-overc as ansible was originally used to
configure things in the OverC framework. The use of ansible was
dropped, however, due to size concerns, the limited functionality we
were using, increased boot times and other factors. It no longer makes
sense to host this recipe in meta-overc and thus we are making it
available as part of meta-cloud-services as it is still a useful tool
for cloud deployments.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>