consul: consule-migrate: move from meta-overc
After several requests to make the consul recipe available outside of
meta-overc and the OverC framework we are moving the consul recipes
here. These are a copy from meta-overc (HEAD 6afe8d2c12ac).

NOTE that not all the DEPENDS have been copied as we will follow this
commit with work to make use of SRC_URIs for dependencies. See the
associated commit(s) which will follow this one.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>