consul-migrate: use SRC_URIs not DEPENDS
It is common to have two GO packages, A and B, and they both depend on
a GO dependency, C. This can be satisfied using the current technique
of writing recipes for each dependency.

BUT it is also common for GO packages to not only specify other
packages in their GoDeps but also specific dependency versions. In the
case that A and B share a dependency on C, but require different
versions, there is no easy way to accomplish this using recipes.

Since GO dependencies are also source dependencies and these
dependencies are overwhelmingly useless on their own, writing recipes
for each is overkill.

We therefor replace DEPENDS with SRC_URIs and SRCREVs as a mechanism
to provide build dependencies.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>