consul-migrate: update to use go.bbclass
We do not want meta-cloud-services to depend on meta-overc so we
switch to use the go.bbclass which is now part of oe-core/yocto. We
also capture all the licenses for all source packages which are
included in consul-migrate and adjust paths to account for the
different handling of go.bbclass vs. golang.bbclass.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>