consul: don't build with the '-linkshared' option
We are currently seeing the following error:
type..eUfRrckI: missing section for relocation target type..FEeY9FdU
...

This is similar to the reported issues
http://www.lyddzz.com/github_/golang/go/issues/30768
and
https://github.com/hashicorp/consul/issues/5486 (but this one appears
to be a report of 'our' issue so isn't really valid)

The go.bbclass currently completes 2 builds (if GO_DYNLINK is set),
one without '-linkshared' and one with. The '--linkshared' option is
marked as experimental (https://golang.org/cmd/link/ -- "-linkshared
Link against installed Go shared libraries (experimental)").

By setting GO_LINKSHARED prior to inheriting go.bbclass we can
overwrite the default compile behavior and only perform the build
without the '-linkshared' option. This shouldn't change the
functionality of consul and we can revert to the default behavior when
the issues have been addressed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>